-- First Commit -- Base score, base game manager, base score manager, base scriptable object events workflow.
- Shooter
Playable with mouse. Need the headset to implement vr. And is missing some bookends
Missing bookends, like replay etc.
Done:
Shooter manager, with events to when to flip targets, when to flip back and when to flip back.
Ui with time left and score. Using scriptableObject events. Wonder if needs to be a Uimanager singleton?
Game ends after 60.
Target flipper responsible to flip the target and flip it back.
Base Target Score. Has a event to when is shot that the target flipper listens to flip back.
Missing:
Target score is missing the formula from the gdd.

- Skee ball
Unplayable but base logic done, have to continue with the vr headset to implement  grabbing a ball and throwing it.
And a lot of bookend is still missing.